added test for MockFile.WriteAllBytes fixed behavior of MockFile.WriteAllBytes
