+++ TestHelpers.Tests/TestHelpers.Tests.csproj
@@ -99,6 +99,7 @@
     <Compile Include="MockFileStreamTests.cs" />
     <Compile Include="MockFileSystemTests.cs" />
     <Compile Include="MockFileTests.cs" />
+    <Compile Include="MockFileWriteAllBytesTests.cs" />
     <Compile Include="MockFileWriteAllTextTests.cs" />
     <Compile Include="MockPathTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
